+++ R/fetch.R
@@ -1,10 +1,10 @@
 #' Fetch Westminster parliamentary constituencies
 #'
 #' Fetch a data frame of all Westminster Parliamentary Constituencies for a
-#' given year and country based on the dfeR::wd_pcon_lad_la_rgn_ctry file
+#' given year and country based on the dfeR::geo_hierarchy file.
 #'
 #' @param year year to filter the locations to, default is "All",
-#' options of 2017, 2019, 2020, 2021, 2022", 2023, 2024
+#' options of 2017, 2019, 2020, 2021, 2022", 2023, 2024, 2025
 #' @param countries vector of desired countries to filter the locations to,
 #' default is "All", or can be a vector with options of "England", "Scotland",
 #' "Wales" or "Northern Ireland"
@@ -26,6 +26,8 @@
 #'
 #' head(fetch_pcons(year = 2023, countries = c("England", "Wales")))
 #'
+#' head(fetch_mayoral())
+#'
 #' fetch_lads(2024, "Wales")
 #'
 #' fetch_las(2022, "Northern Ireland")
@@ -39,7 +41,7 @@ fetch_pcons <- function(year = "All", countries = "All") {
 
   # Helper function to filter to locations we want
   output <- fetch_locations(
-    lookup_data = dfeR::wd_pcon_lad_la_rgn_ctry,
+    lookup_data = dfeR::geo_hierarchy,
     cols = c("pcon_code", "pcon_name"),
     year = year,
     countries = countries
@@ -49,6 +51,9 @@ fetch_pcons <- function(year = "All", countries = "All") {
 }
 
 #' Fetch local authority districts
+#'
+#' Fetch a data frame of all local authority districts for a
+#' given year and country based on the dfeR::geo_hierarchy file.
 #'
 #' @inheritParams fetch
 #'
@@ -63,7 +68,7 @@ fetch_lads <- function(year = "All", countries = "All") {
 
   # Helper function to filter to locations we want
   output <- fetch_locations(
-    lookup_data = dfeR::wd_pcon_lad_la_rgn_ctry,
+    lookup_data = dfeR::geo_hierarchy,
     cols = c("lad_code", "lad_name"),
     year = year,
     countries = countries
@@ -73,6 +78,9 @@ fetch_lads <- function(year = "All", countries = "All") {
 }
 
 #' Fetch local authorities
+#'
+#' Fetch a data frame of all local authorities for a given year and country
+#' based on the dfeR::geo_hierarchy file.
 #'
 #' @inheritParams fetch
 #'
@@ -87,8 +95,8 @@ fetch_las <- function(year = "All", countries = "All") {
 
   # Helper function to filter to locations we want
   output <- fetch_locations(
-    lookup_data = dfeR::wd_pcon_lad_la_rgn_ctry,
-    cols = c("new_la_code", "la_name"),
+    lookup_data = dfeR::geo_hierarchy,
+    cols = c("new_la_code", "la_name", "old_la_code"),
     year = year,
     countries = countries
   )
@@ -97,6 +105,9 @@ fetch_las <- function(year = "All", countries = "All") {
 }
 
 #' Fetch wards
+#'
+#' Fetch a data frame of all wards for a given year and country based on the
+#' dfeR::geo_hierarchy file.
 #'
 #' @inheritParams fetch
 #'
@@ -111,7 +122,7 @@ fetch_wards <- function(year = "All", countries = "All") {
 
   # Helper function to filter to locations we want
   output <- fetch_locations(
-    lookup_data = dfeR::wd_pcon_lad_la_rgn_ctry,
+    lookup_data = dfeR::geo_hierarchy,
     cols = c("ward_code", "ward_name"),
     year = year,
     countries = countries
@@ -120,8 +131,45 @@ fetch_wards <- function(year = "All", countries = "All") {
   return(output)
 }
 
+#' Fetch mayoral combined authorities
+#'
+#' Fetch a data frame of all mayoral combined authorities for a given year
+#' and country based on the dfeR::geo_hierarchy file.
+#'
+#' Note that mayoral combined authorities only exist for England.
+#'
+#' Mayoral combined authorities are also known as English Devolved Areas, as
+#' we add in the Greater London Authority to the combined authority lookup
+#' published by ONS.
+#'
+#' @param year year to filter the locations to, default is "All",
+#' options of 2017, 2019, 2020, 2021, 2022", 2023, 2024, 2025
+#'
+#' @family fetch_locations
+#' @return data frame of unique location names and codes
+#' @export
+#'
+#' @inherit fetch examples
+fetch_mayoral <- function(year = "All") {
+  # Helper function to check the inputs are valid (only England for mayoral)
+  check_fetch_location_inputs(year, "England")
+
+  # Helper function to filter to locations we want
+  output <- fetch_locations(
+    lookup_data = dfeR::geo_hierarchy,
+    cols = c("english_devolved_area_code", "english_devolved_area_name"),
+    year = year,
+    countries = "England"
+  ) |>
+    dplyr::arrange("english_devolved_area_code")
+
+  # Drop rows where not applicable
+  return(output[output$english_devolved_area_code != "z", ])
+}
 
 #' Fetch regions
+#'
+#' Fetch a data frame of all regions based on the dfeR::regions file.
 #'
 #' @family fetch_locations
 #' @return data frame of unique location names and codes
@@ -132,8 +180,9 @@ fetch_regions <- function() {
   dfeR::regions
 }
 
-
 #' Fetch countries
+#'
+#' Fetch a data frame of all countries based on the dfeR::countries file.
 #'
 #' @family fetch_locations
 #' @return data frame of unique location names and codes