+++ data-raw/geo_hierarchy.R
@@ -1,20 +1,24 @@
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# THIS DATA IS DEPRECATED - UPDATE geo_hierarchy INSTEAD
+# Create the geo_hierarchy data set
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# NOTE: Other hierarchies are available, but we expect this to be the main one
+# for analysts working in DfE.
 #
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# Create the wd_pcon_lad_la_reg_ctry data set
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # NOTE: API queries in this script may take a couple of minutes to run
 #
 # This data is created by joining wd_pcon_lad_la
 # - https://geoportal.statistics.gov.uk/search?tags=lup_wd_pcon_lad_utla
-# and lad_region from these lookups
+# and lad_region from
 # - https://geoportal.statistics.gov.uk/search?q=lup_wd_lad_cty_rgn_gor_ctry
+# and combined mayoral authorities from
+# - https://geoportal.statistics.gov.uk/search?tags=LUP_LAD_CAUTH
 #
 # We make some small changes to these currently, more details are in the
-# public facing documentation in R/datasets_documentation.R and the functions
-# in R/datasets_utils.R.
+# public facing documentation in R/datasets_documentation.R.
+#
+# A number of the functions used in this script are stored in
+# R/datasets_utils.R, in there you can see the details of how we query the API
+# and any transformations we apply to the data within those functions.
 #
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # To update this data
@@ -42,8 +46,8 @@
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #
 # 1. ONS like to vary their data set ids
-#    You may need edit the `case_when()` in the `get_wd_pcon_lad_la()`
-#    or `get_lad_region()` functions to add a new condition for the latest
+#    You may need edit the `case_when()` in the `get_*()`
+#    functions to add a new condition for the latest
 #    year, both of these are defined in R/datasets_utils.R
 #
 # 2. Watch out for standard column names changing
@@ -52,7 +56,7 @@
 #    years
 #
 # 3. Regions not joining
-#    The two files are published out of sync, so the years aren't a perfect
+#    The source lookups are published out of sync, so the years aren't a perfect
 #    match. It's pretty possible that in the future we may need to do some
 #    manual joins. In this case add them to the bottom section and make sure
 #    to document in R/datasets_documentation.R, as well as giving the source
@@ -61,24 +65,33 @@
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 library(readxl)
 library(dplyr)
+library(tidyr)
 
 # Set the years we want to pull for -------------------------------------------
 # Started publishing in 2017, but didn't publish a 2018 file
-wd_pcon_lad_la_years <- c(2017, 2019:2024)
+wd_pcon_lad_la_years <- c(2017, 2019:2025)
 
 # Skipping 2021 as not published in API for lad_region and we have enough
 # coverage from other years to join without any gaps
-lad_region_years <- c(2017:2020, 2022:2023)
+lad_region_years <- c(2017:2020, 2022:2025)
+
+mayoral_years <- c(2017:2025)
 
 # Get the main lookup ---------------------------------------------------------
-# Functions defined in R/datasets_utils.R
 # Start with a list of individual data frames per year
 wd_pcon_lad_la <- lapply(wd_pcon_lad_la_years, get_wd_pcon_lad_la) |>
   create_time_series_lookup() |> # smush together into single data frame
   dplyr::select(
-    "first_available_year_included", "most_recent_year_included",
-    "ward_name", "pcon_name", "lad_name", "la_name",
-    "ward_code", "pcon_code", "lad_code", "new_la_code",
+    "first_available_year_included",
+    "most_recent_year_included",
+    "ward_name",
+    "pcon_name",
+    "lad_name",
+    "la_name",
+    "ward_code",
+    "pcon_code",
+    "lad_code",
+    "new_la_code",
   )
 
 # Get the regions lookup ------------------------------------------------------
@@ -89,12 +102,12 @@ regions_lookup <- lapply(lad_region_years, get_lad_region) |>
   mutate(
     region_name = if_else(grepl("^S", lad_code), "Scotland", region_name),
     region_code = if_else(grepl("^S", lad_code), "S92000003", region_code)
-  ) %>%
+  ) |>
   # Get just the unique values to join with
   dplyr::distinct(lad_name, lad_code, region_name, region_code)
 
 # Join on regions -------------------------------------------------------------
-wd_pcon_lad_la_rgn <- wd_pcon_lad_la %>%
+wd_pcon_lad_la_rgn <- wd_pcon_lad_la |>
   dplyr::left_join(
     # Join all rows that aren't missing a region name
     na.omit(regions_lookup, cols = "region_name"),
@@ -102,7 +115,7 @@ wd_pcon_lad_la_rgn <- wd_pcon_lad_la %>%
   )
 
 # Join on countries -----------------------------------------------------------
-wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn %>%
+wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn |>
   dplyr::mutate(
     "country_name" = dplyr::case_when(
       startsWith(pcon_code, "E") ~ "England",
@@ -120,41 +133,116 @@ wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn %>%
     )
   )
 
+# Join on combined mayoral authorities ----------------------------------------
+# NOTE: Some LADs have had different combined authorities over time
+cauth <- lapply(mayoral_years, get_cauth_lad) |>
+  create_time_series_lookup() |>
+  dplyr::select(
+    "first_available_year_included",
+    "most_recent_year_included",
+    "lad_name",
+    "lad_code",
+    "cauth_name",
+    "cauth_code"
+  ) |>
+  dplyr::distinct()
+
+# Join mayoral columns using exploded tables
+wd_pcon_lad_la_cauth_rgn_ctry <- explode_timeseries(wd_pcon_lad_la_rgn_ctry) |>
+  dplyr::left_join(
+    explode_timeseries(cauth),
+    by = c(
+      "lad_code" = "lad_code",
+      "lad_name" = "lad_name",
+      "year" = "year"
+    )
+  ) |>
+  collapse_timeseries() # return to abridged time series
+
+# Fill in blank combined authority names and codes
+# Using z and Not applicable
+geo_hierarchy <- wd_pcon_lad_la_cauth_rgn_ctry |>
+  dplyr::mutate(
+    cauth_name = dplyr::if_else(
+      is.na(cauth_name),
+      "Not applicable",
+      cauth_name
+    ),
+    cauth_code = dplyr::if_else(is.na(cauth_code), "z", cauth_code)
+  )
+
 # Manual fixes ----------------------------------------------------------------
 # !IMPORTANT! Make sure to log all of these in the description for the file in
 # the `R/datasets_documentation.R` script
-wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn_ctry %>%
+geo_hierarchy <- geo_hierarchy |>
   # ONS seemed to miss a 0 in 2017 for Glasgow East PCon
   mutate(across(everything(), ~ ifelse(. == "S1400030", "S14000030", .)))
 
-# Add 3 digit local authority codes from GIAS  ----------------------------
-wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn_ctry %>%
+# Add in Greater London Authority, and convert to english devolved area -------
+geo_hierarchy <- geo_hierarchy |>
+  # Add in GLA based on LAD being London borough
+  dplyr::mutate(
+    cauth_name = dplyr::if_else(
+      stringr::str_starts(lad_code, "E090"),
+      "Greater London Authority",
+      cauth_name
+    ),
+    cauth_code = dplyr::if_else(
+      stringr::str_starts(lad_code, "E090"),
+      "E61000001",
+      cauth_code
+    )
+  ) |>
+  dplyr::rename(
+    "english_devolved_area_name" = "cauth_name",
+    "english_devolved_area_code" = "cauth_code"
+  )
+
+# Add 3 digit local authority codes from GIAS  --------------------------------
+geo_hierarchy <- geo_hierarchy |>
   # join the data onto the GIAs LA 3 digit code data
-  dplyr::left_join(old_la_codes, by = c(
-    "la_name" = "la_name",
-    "new_la_code" = "new_la_code"
-  )) %>%
-  dplyr::mutate(old_la_code = dplyr::if_else(is.na(old_la_code),
-    "z", old_la_code
-  )) %>%
+  dplyr::left_join(
+    old_la_codes,
+    by = c(
+      "la_name" = "la_name",
+      "new_la_code" = "new_la_code"
+    )
+  ) |>
+  dplyr::mutate(
+    old_la_code = dplyr::if_else(is.na(old_la_code), "z", old_la_code)
+  ) |>
   dplyr::distinct()
 
 # Set the order of the columns ------------------------------------------------
-wd_pcon_lad_la_rgn_ctry <- wd_pcon_lad_la_rgn_ctry %>%
+geo_hierarchy <- geo_hierarchy |>
   dplyr::select(
-    "first_available_year_included", "most_recent_year_included",
-    "ward_name", "pcon_name", "lad_name", "la_name",
-    "region_name", "country_name",
-    "ward_code", "pcon_code", "lad_code", "old_la_code", "new_la_code",
-    "region_code", "country_code"
+    "first_available_year_included",
+    "most_recent_year_included",
+    "ward_name",
+    "pcon_name",
+    "lad_name",
+    "la_name",
+    "english_devolved_area_name",
+    "region_name",
+    "country_name",
+    "ward_code",
+    "pcon_code",
+    "lad_code",
+    "old_la_code",
+    "new_la_code",
+    "english_devolved_area_code",
+    "region_code",
+    "country_code"
   )
 
 # QA the joining --------------------------------------------------------------
 # Check for any regions that failed to join
-region_error_check <- wd_pcon_lad_la_rgn_ctry %>%
-  filter(
-    region_code == "" | region_name == "" |
-      is.na(region_name) | is.na(region_code)
+region_error_check <- geo_hierarchy |>
+  dplyr::filter(
+    region_code == "" |
+      region_name == "" |
+      is.na(region_name) |
+      is.na(region_code)
   )
 
 if (nrow(region_error_check) > 0) {
@@ -162,16 +250,12 @@ if (nrow(region_error_check) > 0) {
 }
 
 # Check for any countries that failed to join
-country_error_check <- wd_pcon_lad_la_rgn_ctry %>%
-  filter(country_code == "ERROR" | country_name == "ERROR")
+country_error_check <- geo_hierarchy |>
+  dplyr::filter(country_code == "ERROR" | country_name == "ERROR")
 
 if (nrow(country_error_check) > 0) {
   stop("Country information failing to match, check why this is happening")
 }
 
 # Write the data into the package ---------------------------------------------
-if (TRUE) {
-  stop("wd_pcon_lad_la_rgn_ctry is deprecated - update geo_hierarchy instead")
-} else {
-  usethis::use_data(wd_pcon_lad_la_rgn_ctry, overwrite = TRUE)
-}
+usethis::use_data(geo_hierarchy, overwrite = TRUE)